+++ internal/models/alert_test.go
@@ -1,10 +1,12 @@
 package models_test
 
 import (
+	"fmt"
 	"sort"
 	"testing"
 
 	"github.com/google/go-cmp/cmp"
+	"github.com/prymitive/karma/internal/config"
 	"github.com/prymitive/karma/internal/models"
 )
 
@@ -41,13 +43,15 @@ func TestLabelsSet(t *testing.T) {
 }
 
 type sortLabelsTestCase struct {
-	in  models.Labels
-	out models.Labels
+	order []string
+	in    models.Labels
+	out   models.Labels
 }
 
 func TestSortLabels(t *testing.T) {
 	testCases := []sortLabelsTestCase{
 		{
+			order: []string{},
 			in: models.Labels{
 				{Name: "foo", Value: "bar"},
 			},
@@ -56,6 +60,7 @@ func TestSortLabels(t *testing.T) {
 			},
 		},
 		{
+			order: []string{},
 			in: models.Labels{
 				{Name: "foo", Value: "bar"},
 				{Name: "bar", Value: "foo"},
@@ -66,6 +71,7 @@ func TestSortLabels(t *testing.T) {
 			},
 		},
 		{
+			order: []string{},
 			in: models.Labels{
 				{Name: "bar", Value: "foo"},
 				{Name: "foo", Value: "bar"},
@@ -76,6 +82,7 @@ func TestSortLabels(t *testing.T) {
 			},
 		},
 		{
+			order: []string{},
 			in: models.Labels{
 				{Name: "foo", Value: "foo"},
 				{Name: "bar", Value: "foo"},
@@ -88,6 +95,7 @@ func TestSortLabels(t *testing.T) {
 			},
 		},
 		{
+			order: []string{},
 			in: models.Labels{
 				{Name: "1", Value: "a12"},
 				{Name: "1", Value: "1"},
@@ -97,15 +105,48 @@ func TestSortLabels(t *testing.T) {
 				{Name: "1", Value: "1"},
 				{Name: "1", Value: "a2"},
 				{Name: "1", Value: "a12"},
+			},
+		},
+		{
+			order: []string{"bar"},
+			in: models.Labels{
+				{Name: "baz", Value: "1"},
+				{Name: "bar", Value: "1"},
+				{Name: "foo", Value: "1"},
+			},
+			out: models.Labels{
+				{Name: "bar", Value: "1"},
+				{Name: "baz", Value: "1"},
+				{Name: "foo", Value: "1"},
+			},
+		},
+		{
+			order: []string{"foo", "bar"},
+			in: models.Labels{
+				{Name: "foo", Value: "a10"},
+				{Name: "bar", Value: "1"},
+				{Name: "foo", Value: "a3"},
+			},
+			out: models.Labels{
+				{Name: "foo", Value: "a3"},
+				{Name: "foo", Value: "a10"},
+				{Name: "bar", Value: "1"},
 			},
 		},
 	}
 
-	for _, testCase := range testCases {
-		sort.Sort(testCase.in)
-		if diff := cmp.Diff(testCase.in, testCase.out); diff != "" {
-			t.Errorf("Incorrectly sorted labels (-want +got):\n%s", diff)
-			break
-		}
+	defer func() {
+		config.Config.Labels.Order = []string{}
+	}()
+
+	for i, testCase := range testCases {
+		t.Run(fmt.Sprintf("[%d] order=%v", i, testCase.order), func(t *testing.T) {
+			config.Config.Labels.Order = testCase.order
+			sort.Sort(testCase.in)
+			if diff := cmp.Diff(testCase.in, testCase.out); diff != "" {
+				t.Errorf("Incorrectly sorted labels (-want +got):\n%s", diff)
+				t.FailNow()
+			}
+		})
 	}
 }